+++ sdk/storage/azure_storage_blob/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.1.0-beta.2 (Unreleased)
+## 1.1.0-beta.3 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,16 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.1.0-beta.2 (2026-08-11)
+
+### Features Added
+
+- Added `BlobClient::start_copy_from_url()` and `BlobClient::abort_copy()` for asynchronous blob copy operations.
+
+### Breaking Changes
+
+- Removed `BlobContainerClient::list_blobs_hierarchical()`.
 
 ## 1.1.0-beta.1 (2026-07-14)